Fixed SQL to get JC Decaux station list
This was not correctly updating the database, but now seems to be
working, subject to further testing.  It uses an 'upsert' logic, i.e. if
a record with the same contract name and station number exists, update
its attributes with the latest data, and if no such record exists,
insert it.